+++ tests/functional/test_zc.py
@@ -21,4 +21,10 @@ from NeoVintageous.tests import unittest
 class Test_zc(unittest.FunctionalTestCase):
 
     def test_n(self):
-        self.eq('x\nfi|zz x', 'n_zc', 'x\n|fizz x')
+        for seq in ('zc', 'zC'):
+            self.eq('x\nfi|zz x', 'n_' + seq, 'x\n|fizz x')
+
+    def test_n_html(self):
+        self.syntax('Packages/HTML/HTML.sublime-syntax')
+        self.eq('<div>\n|    test\n</div>\n', 'n_zc', '|<div>\n    test\n</div>\n')
+        self.assertFolded('<div>|\n    test|\n</div>\n')